fix: 500kb+ -> 194kb
