+++ package.json
@@ -1,5 +1,6 @@
 {
-  "name": "@blockstack/explorer",
+  "name": "@hiro/explorer",
+  "description": "The source code for the Stacks Blockchain explorer.",
   "scripts": {
     "build": "next telemetry disable && next build",
     "build:analyze": "next telemetry disable && ANALYZE=true next build",
@@ -13,142 +14,101 @@
     "typecheck": "tsc --noEmit"
   },
   "version": "1.0.0",
+  "author": "Hiro Systems PBC (formerly Blockstack PBC)",
   "dependencies": {
-    "@blockstack/stacks-blockchain-api-types": "^0.44.0",
-    "@emotion/babel-plugin": "^11.1.2",
-    "@emotion/babel-preset-css-prop": "^11.0.0",
-    "@emotion/cache": "^11.0.0",
+    "@babel/core": "^7.12.10",
     "@emotion/core": "^11.0.0",
     "@emotion/css": "^11.0.0",
     "@emotion/react": "^11.1.4",
     "@emotion/server": "^11.0.0",
     "@promster/express": "^5.0.0",
     "@promster/server": "^6.0.0",
-    "@reach/tooltip": "^0.12.1",
     "@react-aria/interactions": "^3.3.2",
     "@react-spectrum/utils": "^3.5.0",
-    "@react-stately/toggle": "^3.2.1",
     "@stacks/auth": "^1.0.1",
     "@stacks/connect": "^4.3.16-alpha.d514fd4.0",
-    "@stacks/connect-ui": "^2.17.17-alpha.d514fd4.0",
     "@stacks/network": "^1.0.1",
     "@stacks/transactions": "^1.0.1",
     "@stacks/ui": "7.3.1-beta.2",
     "@stacks/ui-core": "7.0.0",
     "@stacks/ui-theme": "7.1.1-beta.1",
     "@stacks/ui-utils": "7.1.1-beta.1",
-    "@styled-system/theme-get": "^5.1.2",
-    "@tabler/icons": "^1.38.1",
+    "@tabler/icons": "^1.39.1",
     "@tippyjs/react": "^4.2.0",
-    "@types/just-debounce-it": "^1.1.0",
-    "@types/nprogress": "^0.2.0",
-    "@types/pluralize": "^0.0.29",
-    "@types/prismjs": "^1.16.1",
     "@vkontakte/vk-qr": "^2.0.10",
-    "@welldone-software/why-did-you-render": "^6.0.5",
-    "ansi-to-react": "^6.1.4",
-    "as-bind": "^0.5.0",
     "awesome-debounce-promise": "^2.1.0",
-    "babel-plugin-module-resolver": "^4.1.0",
     "bignumber.js": "^9.0.1",
-    "blockstack": "^21.1.1",
     "bn.js": "^5.1.3",
     "brotli-webpack-plugin": "^1.1.0",
+    "buffer": "^6.0.3",
     "c32check": "^1.1.2",
-    "cacheable-response": "^2.1.9",
     "compression": "^1.7.4",
     "dayjs": "^1.10.3",
-    "dequal": "^2.0.2",
     "devalue": "^2.0.1",
-    "elliptic": "^6.5.3",
     "express": "^4.17.1",
     "express-static-gzip": "^2.1.1",
-    "extract-json-string": "^1.0.1",
     "fathom-client": "^3.0.0",
-    "fontfaceobserver": "^2.1.0",
     "formik": "^2.2.6",
     "just-debounce-it": "^1.1.0",
     "lru-cache": "^6.0.0",
     "mdi-react": "^7.3.0",
     "modern-normalize": "^1.0.0",
     "morgan": "^1.10.0",
-    "next": "^10.0.6-canary.6",
+    "next": "^10.0.5",
     "next-plugin-preact": "^3.0.3",
-    "noble-secp256k1": "^1.1.1",
-    "node-zopfli-es": "^1.0.7",
     "nookies": "^2.5.1",
     "nprogress": "^0.2.0",
     "pluralize": "^8.0.0",
-    "preact": "^10.5.10",
-    "preact-context-provider": "^1.2.1",
+    "preact": "^10.5.11",
     "preact-render-to-string": "^5.1.12",
-    "prism-react-renderer": "^1.1.1",
     "prismjs": "^1.23.0",
-    "prom-client": "^13.0.0",
-    "react": "^17.0.1",
+    "react": "npm:@preact/compat",
     "react-aria": "^3.3.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "npm:@preact/compat",
     "react-error-boundary": "^3.1.0",
-    "react-fast-compare": "^3.2.0",
-    "react-hooks-sse": "^2.0.0",
     "react-query": "^3.5.16",
     "react-simple-code-editor": "^0.11.0",
-    "react-ssr-prepass": "^1.2.1",
+    "react-ssr-prepass": "npm:preact-ssr-prepass",
     "react-use": "^15.3.8",
-    "react-virtuoso": "^1.3.0",
-    "readable-stream": "^3.6.0",
     "recoil": "facebookexperimental/Recoil.git#nightly",
-    "shrink-ray-current": "^4.1.2",
-    "store": "^2.0.12",
-    "strip-ansi": "^6.0.0",
     "swr": "^0.4.0",
-    "tweakpane": "^1.5.7",
     "unique-names-generator": "^4.3.1",
     "use-constant": "^1.0.0",
     "use-events": "^1.4.2",
-    "use-onclickoutside": "^0.3.1",
-    "valtio": "^0.6.1",
-    "wasm-loader": "^1.3.0",
     "web-api-hooks": "^3.0.2",
-    "webassembly": "^0.11.0",
-    "webpack": "^5.16.0",
+    "webpack": "4.46.0",
     "yup": "^0.32.8"
   },
   "devDependencies": {
-    "@blockstack/stacks-blockchain-sidecar-types": "^0.0.22",
+    "@blockstack/stacks-blockchain-api-types": "^0.44.0",
     "@commitlint/config-conventional": "^11.0.0",
+    "@emotion/babel-plugin": "^11.1.2",
+    "@emotion/babel-preset-css-prop": "^11.0.0",
     "@next/bundle-analyzer": "^10.0.5",
     "@stacks/eslint-config": "^1.0.7",
     "@stacks/prettier-config": "^0.0.7",
-    "@types/express": "^4.17.11",
-    "@types/fontfaceobserver": "^0.0.6",
-    "@types/next": "^9.0.0",
-    "@types/node": "^14.14.22",
+    "@types/just-debounce-it": "^1.1.0",
+    "@types/nprogress": "^0.2.0",
+    "@types/pluralize": "^0.0.29",
+    "@types/prismjs": "^1.16.1",
     "@types/react": "^17.0.0",
-    "@types/store": "^2.0.2",
-    "@types/styled-system__theme-get": "^5.0.1",
-    "@typescript-eslint/eslint-plugin": "^4.14.0",
     "@typescript-eslint/parser": "^4.14.0",
+    "@welldone-software/why-did-you-render": "^6.0.5",
     "commitlint": "^11.0.0",
     "eslint": "^7.18.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-prettier": "^3.3.1",
     "husky": "^4.3.8",
-    "lint-staged": "^10.5.3",
-    "package-json-cleanup-loader": "^1.0.3",
     "prettier": "^2.2.1",
     "typescript": "^4.1.3"
   },
-  "resolutions": {
-    "eslint": "^7.16.0",
-    "bn.js": "^5.1.3"
-  },
+  "homepage": "https://explorer.stacks.co",
   "husky": {
     "hooks": {
       "pre-commit": "yarn lint:prettier:fix && yarn typecheck",
       "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
     }
   },
+  "license": "MIT",
+  "prettier": "@stacks/prettier-config",
   "release": {
     "plugins": [
       "@semantic-release/commit-analyzer",
@@ -166,6 +126,10 @@
       ]
     ]
   },
-  "license": "MIT",
-  "prettier": "@stacks/prettier-config"
+  "resolutions": {
+    "eslint": "^7.16.0",
+    "bn.js": "^5.1.3",
+    "buffer": "^6.0.3",
+    "webpack": "4.46.0"
+  }
 }
