+++ src/common/hooks/use-filter-state.tsx
@@ -1,8 +1,8 @@
 import { useCallback } from 'react';
 import { useRecoilCallback, useRecoilState } from 'recoil';
 
-import type { TxTypeFilterOptions } from '@store/sandbox';
-import { filterState } from '@store/sandbox';
+import type { TxTypeFilterOptions } from '@store/filter';
+import { filterState } from '@store/filter';
 
 export const useFilterState = (key: 'sandbox' | 'txList', showCoinbase = false) => {
   const [filter, setFilterState] = useRecoilState(filterState(key));
