--- src/common/sandbox/contracts/index.ts
@@ -1,4 +1,0 @@
-export * from './hello-world-contract';
-export * from './kv-store';
-export * from './status';
-export * from './stream';
