--- src/common/sandbox/contracts/kv-store.tsx
@@ -1,16 +1,0 @@
-export const kvStoreContract = {
-  name: 'kv-store',
-  source: `;; kv-store contract
-
-(define-map store ((key (buff 32))) ((value (buff 32))))
-
-(define-public (get-value (key (buff 32)))
-    (match (map-get? store {key: key})
-        entry (ok (get value entry))
-        (err 0)))
-
-(define-public (set-value (key (buff 32)) (value (buff 32)))
-    (begin
-        (map-set store {key: key} {value: value})
-        (ok u1)))`,
-};
