--- src/common/sandbox/contracts/status.tsx
@@ -1,36 +1,0 @@
-export const statusContract = {
-  name: 'status',
-  source: `;; status contract
-
-(define-map statuses
-  (
-    (author principal)
-  )
-  (
-    (status (buff 512))
-  )
-)
-
-(define-read-only (get-status (author principal))
-  (begin
-    (print author)
-    (default-to ""
-      (get status (map-get? statuses {author: author}))
-    )
-  )
-)
-
-(define-public (write-status!
-    (status (buff 512))
-  )
-  (begin
-    (print tx-sender)
-    (print status)
-    (map-set statuses
-      ((author tx-sender))
-      ((status status))
-    )
-    (ok status)
-  )
-)`,
-};
