--- src/common/sandbox/examples.ts
@@ -1,11 +1,0 @@
-import {
-  helloWorldContract,
-  kvStoreContract,
-  statusContract,
-  streamContract,
-} from '@common/sandbox/contracts';
-
-export const SampleContracts: readonly {
-  readonly name: string;
-  readonly source: string;
-}[] = [helloWorldContract, kvStoreContract, statusContract, streamContract];
