--- src/components/code-editor/index.tsx
@@ -1,3 +1,0 @@
-import dynamic from 'next/dynamic';
-
-export const CodeEditor = dynamic(() => import('./code-editor'));
