+++ src/pages/_document.tsx
@@ -59,7 +59,6 @@ export default class MyDocument extends Document<DocumentProps> {
 })();`,
             }}
           />
-
           <link rel="preconnect" href="https://cdn.usefathom.com" crossOrigin="true" />
           <link rel="preconnect" href="https://fonts.googleapis.com" crossOrigin="true" />
           <link
