+++ src/store/auth.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const authResponseState = atom<null | string>({
+  key: 'authResponse',
+  default: null,
+});
